# API-Docs/swagger.yaml
openapi: 3.0.1
info:
  title: ProcesadorEnviosAPI
  version: v1
servers:
  - url: /Prod
paths:
  /api/Envios:
    get:
      tags:
        - Envios
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Envio'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Envio'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Envio'
    post:
      tags:
        - Envios
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Envio'
          text/json:
            schema:
              $ref: '#/components/schemas/Envio'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Envio'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Envio'
            application/json:
              schema:
                $ref: '#/components/schemas/Envio'
            text/json:
              schema:
                $ref: '#/components/schemas/Envio'
  '/api/Envios/{envioId}':
    get:
      tags:
        - Envios
      parameters:
        - name: envioId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Envio'
            application/json:
              schema:
                $ref: '#/components/schemas/Envio'
            text/json:
              schema:
                $ref: '#/components/schemas/Envio'
  '/api/Envios/{envioId}/novedades':
    post:
      tags:
        - Envios
      parameters:
        - name: envioId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Novedades'
          text/json:
            schema:
              $ref: '#/components/schemas/Novedades'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Novedades'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Novedades'
            application/json:
              schema:
                $ref: '#/components/schemas/Novedades'
            text/json:
              schema:
                $ref: '#/components/schemas/Novedades'
...
components:
  schemas:
    Envio:
      type: object
      properties:
        id:
          type: integer
          format: int32
        direccionOrigen:
          type: string
          nullable: true
        direccionDestino:
          type: string
          nullable: true
        contactoComprador:
          type: string
          nullable: true
        estadoEnvio:
          type: string
          nullable: true
        detalleProducto:
          type: string
          nullable: true
        operadorLogistico:
          type: integer
          format: int32
      additionalProperties: false
    Novedades:
      type: object
      properties:
        id:
          type: integer
          format: int32
        idEnvio:
          type: integer
          format: int32
        nuevoEstado:
          type: string
          nullable: true
      additionalProperties: false
...